refactor(doc-search): split 583-line index.ts into modules
Extract focused modules by concern:
- types.ts: Document, SearchResult, Metadata, AccessEvent types
- db/init.ts: Database init, assertDb, validateDocumentRow
- db/config.ts: Dynamic config loading from DB
- log/access.ts: JSONL access logging with degraded mode
- render/*.ts: HTML utilities, home, search, doc, not-found pages
- query/tokenizer.ts: normalize, tokenize, buildMatchQuery

Rewrite index.ts as clean entry point:
- Uses createDocSearchApp() factory returning MockAppV2
- Uses parseCliArgs from mock-lib
- Returns { ...mockApp, seed: async () => { ... } } without type assertion
- Uses startServer(app) in import.meta.main guard
